+++ projects/fengine/render/resources/fanMeshManager.cpp
@@ -68,7 +68,7 @@ namespace fan
     //========================================================================================================
     void MeshManager::Remove( const std::string& _path )
     {
-        for( int          meshIndex = 0; meshIndex < mMeshes.size(); meshIndex++ )
+        for( int          meshIndex = 0; meshIndex < (int)mMeshes.size(); meshIndex++ )
         {
             if( mMeshes[meshIndex]->mPath == _path )
             {
